Refactor stage 1
Move everything advantage related inside an advantage folder.

Fixing small bugs I found:
1) get_cutomer_info throws 500 if user_token is missing
2) post_advantage_subscription for already existing product did not
   increment the existing value correctly
